requirements.txt.txt
   discord.py
   anthropic